Refactor logic from controller to service
